requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
anthropic>=0.7.0
python-dotenv>=1.0.0
